refactor : webview mode 추상화
